Decouple Composables from ComponentActivities
This open up the possibility of a single activity app and implementing Jetpack Navigation.
